Refactor Code

+++ app/Http/Controllers/V1/UserController.php
@@ -84,13 +84,13 @@ class UserController extends Controller
             'email'                     => 'required|email|unique:users,email',
             'joining_date'              => 'required|date',
             'ending_date'               => 'nullable|date|after:joining_date',
-            'password'                  => 'required|string|max:8',
-            'address1'                  => 'required|string|max:50',
-            'address2'                  => 'nullable|string|max:50',
+            'password'                  => 'required_if:role_id,1,2,3,7|string|max:8',
+            'address1'                  => 'required|string|max:30',
+            'address2'                  => 'nullable|string|max:30',
             'zip_code'                  => 'required|integer|min:6',
             'phone'                     => 'nullable|integer|min:10',
-            'total_leave'               => 'required|numeric',
-            'used_leave'                => 'nullable|numeric',
+            'total_leave'               => 'required|integer',
+            'used_leave'                => 'nullable|integer',
             'restaurant_id'             => 'required_if:role_id,3,4,5,6|integer|exists:restaurants,id',
             'service_type_id.*'         => 'required_if:role_id,7|integer|exists:service_types,id',
 
@@ -134,17 +134,16 @@ class UserController extends Controller
             'email'                     => 'nullable|email|unique:users,email',
             'joining_date'              => 'nullable|date',
             'ending_date'               => 'nullable|date|after:joining_date',
-            'password'                  => 'nullable|max:8',
-            'address1'                  => 'nullable|string|max:50',
-            'address2'                  => 'nullable|string|max:50',
+            'address1'                  => 'nullable|string|max:30',
+            'address2'                  => 'nullable|string|max:30',
             'zip_code'                  => 'nullable|integer|min:6',
             'phone'                     => 'nullable|integer|min:10',
-            'total_leave'               => 'nullable|numeric',
-            'used_leave'                => 'nullable|numeric',
+            'total_leave'               => 'nullable|integer',
+            'used_leave'                => 'nullable|integer',
         ]);
 
         $user = User::findOrFail($id);
-        $user->update($request->only('role_id', 'first_name', 'last_name', 'email', 'joining_date', 'ending_date', 'address1', 'address2', 'phone', 'total_leave', 'used_leave', 'zip_code') + ['password' => Hash::make($request->password)]);
+        $user->update($request->only('role_id', 'first_name', 'last_name', 'email', 'joining_date', 'ending_date', 'address1', 'address2', 'phone', 'total_leave', 'used_leave', 'zip_code'));
 
         return ok('User updated successfully!', $user);
     }
@@ -171,8 +170,9 @@ class UserController extends Controller
     public function delete($id)
     {
         $user = User::findOrFail($id);
-        if ($user->restaurantUsers()->count() > 0) {
+        if ($user->restaurantUsers()->count() > 0 && $user->vendors()->count() > 0) {
             $user->restaurantUsers()->detach();
+            $user->vendors()->delete();
         }
         $user->delete();
         return ok('User deleted successfully');
